add Makefile, rewrite README, fix IntroAnimation build errors, complete phase 2 UAT
- Makefile wrapping CMake with make/debug/clean/rebuild/run/test/asan targets
- Comprehensive README with dependency table, platform-specific install
  instructions (Linux/macOS/Windows, x64/aarch64), build commands, project
  layout, shader architecture docs
- Fix IntroAnimation.cpp: add missing <cstdint> include, replace
  std::fabsf with std::fabs for GCC 13 compatibility
- Phase 02 UAT: 6/7 passed, 1 skipped (torch not in env), 0 issues

+++ src/intro/IntroAnimation.cpp
@@ -1,6 +1,7 @@
 #include "intro/IntroAnimation.hpp"
 #include <imgui.h>
 #include <cmath>
+#include <cstdint>
 #include <cstdlib>
 #include <algorithm>
 #include <numeric>
@@ -384,7 +385,7 @@ float IntroAnimation::getUIAlpha() const {
 void IntroAnimation::syncBorderToWindow(float x, float y, float w, float h) {
     const float dx = x - m_borderX;
     const float dy = y - m_borderY;
-    if (std::fabsf(dx) < 0.5f && std::fabsf(dy) < 0.5f) return; // already aligned
+    if (std::fabs(dx) < 0.5f && std::fabs(dy) < 0.5f) return; // already aligned
 
     for (auto& p : m_particles) {
         if (!p.goToBorder) continue;
